+++ drivers/ec2/src/main/java/de/uniulm/omi/cloudiator/sword/drivers/profitbricks/EC2JCloudsViewFactory.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package de.uniulm.omi.cloudiator.sword.drivers.ec2;
+package de.uniulm.omi.cloudiator.sword.drivers.profitbricks;
 
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
