+++ drivers/ec2/src/main/test/java/de/uniulm/omi/cloudiator/sword/drivers/ec2/strategy/Ec2CreateVirtualMachineStrategyTest.java
@@ -1,4 +1,4 @@
-package de.uniulm.omi.cloudiator.sword.drivers.ec2.strategy;
+package de.uniulm.omi.cloudiator.sword.drivers.profitbricks.strategy;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
