+++ drivers/profitbricks/pom.xml
@@ -38,7 +38,7 @@
         <dependency>
             <groupId>org.apache.jclouds.labs</groupId>
             <artifactId>profitbricks-rest</artifactId>
-            <version>${jclouds.version}</version>
+            <version>2.0.3-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.apache.jclouds.driver</groupId>
