+++ service/src/main/java/de/uniulm/omi/cloudiator/sword/service/providers/Providers.java
@@ -24,8 +24,8 @@ import com.google.inject.AbstractModule;
 import de.uniulm.omi.cloudiator.sword.base.BaseComputeService;
 import de.uniulm.omi.cloudiator.sword.domain.PropertiesBuilder;
 import de.uniulm.omi.cloudiator.sword.drivers.azure.config.AzureComputeModule;
-import de.uniulm.omi.cloudiator.sword.drivers.ec2.config.Ec2ComputeModule;
-import de.uniulm.omi.cloudiator.sword.drivers.ec2.config.ProfitbricksComputeModule;
+import de.uniulm.omi.cloudiator.sword.drivers.profitbricks.config.Ec2ComputeModule;
+import de.uniulm.omi.cloudiator.sword.drivers.profitbricks.config.ProfitbricksComputeModule;
 import de.uniulm.omi.cloudiator.sword.drivers.flexiant.config.FlexiantComputeModule;
 import de.uniulm.omi.cloudiator.sword.drivers.google.config.GoogleCloudComputeModule;
 import de.uniulm.omi.cloudiator.sword.drivers.openstack.config.OpenstackComputeModule;
@@ -91,9 +91,9 @@ public class Providers {
         PropertiesBuilder.newBuilder().build()));
     //Profitbricks
     final Set<AbstractModule> profitbricksModules = new HashSet<>();
-    azureModules.add(new ProfitbricksComputeModule());
+    profitbricksModules.add(new ProfitbricksComputeModule());
     registerProvider(
-        new ProviderConfiguration("profitbricks", profitbricksModules, BaseComputeService.class,
+        new ProviderConfiguration("profitbricks-rest", profitbricksModules, BaseComputeService.class,
             PropertiesBuilder.newBuilder().build()));
   }
